app/requirements.txt
Django==3.0.7
gunicorn==20.0.4
psycopg2-binary==2.8.5
requests==2.25.0
prettytable==2.0.0
djangorestframework==3.12.2
simplejson==3.17.2